+++ fastfield_codecs/src/main.rs
@@ -90,7 +90,7 @@ fn bench_ip() {
     {
         let mut data = vec![];
         for dataset in dataset.chunks(500_000) {
-            serialize_u128(|| dataset.iter().cloned(), dataset.len() as u64, &mut data).unwrap();
+            serialize_u128(|| dataset.iter().cloned(), dataset.len() as u32, &mut data).unwrap();
         }
         let compression = data.len() as f64 / (dataset.len() * 16) as f64;
         println!("Compression 50_000 chunks {:.4}", compression);
@@ -103,7 +103,7 @@ fn bench_ip() {
     let mut data = vec![];
     {
         print_time!("creation");
-        serialize_u128(|| dataset.iter().cloned(), dataset.len() as u64, &mut data).unwrap();
+        serialize_u128(|| dataset.iter().cloned(), dataset.len() as u32, &mut data).unwrap();
     }
 
     let compression = data.len() as f64 / (dataset.len() * 16) as f64;
